feat(perusahaan cukai HT + HPTL): pembuatan fitur import excel

+++ database/migrations/2023_12_06_040350_create_sbp_table.php
@@ -27,8 +27,8 @@ return new class extends Migration
                 ->nullOnDelete()
                 ->cascadeOnUpdate();
 
-            $table->double('jumlah');
-            $table->double('tindak_lanjut');
+            $table->decimal('jumlah');
+            $table->decimal('tindak_lanjut');
             $table->date('tanggal_input');
             $table->timestamps();
             $table->softDeletes();
